refactor: split serverinfo into server and roles
